refactor: move util to commons package
