+++ packages/core/src/util/RecurringTaskScheduler/RecurringTaskScheduler.test.ts
@@ -17,11 +17,11 @@
  *
  */
 
+import {advanceJestTimersWithPromise} from '@wireapp/commons/lib/util/testUtils';
+
 import {TimeUtil} from '@wireapp/commons';
 
 import {RecurringTaskScheduler} from './RecurringTaskScheduler';
-
-import {advanceJestTimersWithPromise} from '../../testUtils';
 
 const mockedStore = {
   storage: new Map<string, number>(),
